Enhance UI for Intent detail
