+++ app/components/ResponseList/styles.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const SpinningWrapper = styled.div`
+  height: calc(100vh - 150px);
+`;
+
+export const NoResposne = styled.div`
+  height: calc(100vh - 150px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
