+++ app/containers/IntentManagement/IntentManagement.js
@@ -1,28 +1,11 @@
-import itemManagementHoc from 'hoc/ItemManagementHoc';
 import { connect } from 'react-redux';
-import { actions } from 'reducers/intent';
-import {
-  reselectSorting,
-  getFetchingError,
-} from 'selectors/ticket';
-import { SORT } from 'utils/constants';
-import TicketTable from './IntentTable';
-import { toI18n } from '../../utils/func-utils';
-
-const { TICKET_SORT } = SORT;
+import IntentManagement from '../../components/IntentManagement';
+import { getIntentDetailFromRoute } from '../../selectors/intent';
 
 const mapStateToProps = state => ({
-  errorMsg: getFetchingError(state),
-  currentSorting: reselectSorting(state),
-  title: toI18n('ADMIN_INTENT_TABLE_ALL_INTENT'),
-  sortItems: TICKET_SORT,
+  currentIntent: getIntentDetailFromRoute(state),
 });
-
-const mapDispatchToProps = {
-  handleSort: actions.sortTicket,
-};
 
 export default connect(
   mapStateToProps,
-  mapDispatchToProps,
-)(itemManagementHoc(TicketTable));
+)(IntentManagement);
